xatmi: moved 'jump-context' from server to xatmi, where it belongs
Before: `server` had _service state_ that was only used by `xatmi`

Now: xatmi::internal::context holds state for Jump and header -> used
only by xatmi services.
`server` does not know anything about xatmi.
