+++ middleware/xatmi/makefile.cmk
@@ -31,7 +31,7 @@ lib_xatmi_explicit = make.LinkLibrary( 'bin/casual-xatmi-explicit',
       make.Compile( 'source/xatmi/explicit/buffer.cpp'),
       make.Compile( 'source/xatmi/explicit/service.cpp'),
       make.Compile( 'source/xatmi/internal/server/service.cpp'),
-      make.Compile( 'source/xatmi/internal/header/context.cpp'),
+      make.Compile( 'source/xatmi/internal/context.cpp'),
       make.Compile( 'source/xatmi/internal/code.cpp'),
       make.Compile( 'source/xatmi/internal/log.cpp'),
 	],
